backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
requests>=2.31.0 